Move stuff to cover.const
